Program used to simulate transmitted data
This is the Arduino program that was used to generate a triangle wave that simulated data that was sent over serial from the Arduino to the computer. It formats the data the way the program expects to solve certain issues that were faced. Mainly, if the program started somewhere between the start and end of a frame then the data would not be read correctly, so the first partial frame is read in with the new line character and thrown out. There is also a header to give the name of the data coming in.